+++ R/skip.R
@@ -114,8 +114,14 @@ skip_if_not_installed <- function(pkg, minimum_version = NULL) {
     installed_version <- package_version(pkg)
     if (installed_version < minimum_version) {
       skip(paste0(
-        "Installed ", pkg, " is version ", installed_version, "; ",
-        "but ", minimum_version, " is required"
+        "Installed ",
+        pkg,
+        " is version ",
+        installed_version,
+        "; ",
+        "but ",
+        minimum_version,
+        " is required"
       ))
     }
   }
@@ -125,6 +131,34 @@ skip_if_not_installed <- function(pkg, minimum_version = NULL) {
 package_version <- function(x) {
   utils::packageVersion(x)
 }
+
+#' @export
+#' @param spec A version specification like '>= 4.1.0' denoting that this test
+#'   should only be run on R versions 4.1.0 and later.
+#' @rdname skip
+skip_unless_r <- function(spec) {
+  parts <- unlist(strsplit(spec, " ", fixed = TRUE))
+  if (length(parts) != 2L) {
+    cli::cli_abort(
+      "{.arg spec} should be a comparison like '>=' and an R version separated by a space."
+    )
+  }
+  comparator <- match.fun(parts[1L])
+  required_version <- numeric_version(parts[2L])
+
+  current_version <- getRversion()
+  skip_if_not(
+    comparator(current_version, required_version),
+    sprintf(
+      "Current R version (%s) does not satisfy requirement (%s %s)",
+      current_version,
+      parts[1L],
+      required_version
+    )
+  )
+}
+# for mocking
+getRversion <- NULL
 
 #' @export
 #' @rdname skip
@@ -146,7 +180,8 @@ skip_on_cran <- function() {
 
 #' @export
 #' @param os Character vector of one or more operating systems to skip on.
-#'   Supported values are `"windows"`, `"mac"`, `"linux"`, and `"solaris"`.
+#'   Supported values are `"windows"`, `"mac"`, `"linux"`, `"solaris"`,
+#'   and `"emscripten"`.
 #' @param arch Character vector of one or more architectures to skip on.
 #'   Common values include `"i386"` (32 bit), `"x86_64"` (64 bit), and
 #'   `"aarch64"` (M1 mac). Supplying `arch` makes the test stricter; i.e. both
@@ -155,15 +190,17 @@ skip_on_cran <- function() {
 skip_on_os <- function(os, arch = NULL) {
   os <- match.arg(
     os,
-    choices = c("windows", "mac", "linux", "solaris"),
+    choices = c("windows", "mac", "linux", "solaris", "emscripten"),
     several.ok = TRUE
   )
 
-  msg <- switch(system_os(),
+  msg <- switch(
+    system_os(),
     windows = if ("windows" %in% os) "On Windows",
-    darwin =  if ("mac" %in% os) "On Mac",
-    linux =   if ("linux" %in% os) "On Linux",
-    sunos =   if ("solaris" %in% os) "On Solaris"
+    darwin = if ("mac" %in% os) "On Mac",
+    linux = if ("linux" %in% os) "On Linux",
+    sunos = if ("solaris" %in% os) "On Solaris",
+    emscripten = if ("emscripten" %in% os) "On Emscripten"
   )
 
   if (!is.null(arch) && !is.null(msg)) {
@@ -246,7 +283,7 @@ skip_on_appveyor <- function() {
 # helpers -----------------------------------------------------------------
 
 on_ci <- function() {
- env_var_is_true("CI")
+  env_var_is_true("CI")
 }
 in_covr <- function() {
   env_var_is_true("R_COVR")