+++ R/watcher.R
@@ -28,7 +28,9 @@ watch <- function(path, callback, pattern = NULL, hash = TRUE) {
     if (changes$n > 0) {
       # cat("C")
       keep_going <- TRUE
-      try(keep_going <- callback(changes$added, changes$deleted, changes$modified))
+      try(
+        keep_going <- callback(changes$added, changes$deleted, changes$modified)
+      )
 
       if (!isTRUE(keep_going)) return(invisible())
     } else {
@@ -40,11 +42,17 @@ watch <- function(path, callback, pattern = NULL, hash = TRUE) {
 }
 
 safe_digest <- function(path) {
-  if (!file.exists(path)) return(NA_character_)
-  if (is_directory(path)) return(NA_character_)
-  if (!is_readable(path)) return(NA_character_)
+  if (!file.exists(path)) {
+    return(NA_character_)
+  }
+  if (is_directory(path)) {
+    return(NA_character_)
+  }
+  if (!is_readable(path)) {
+    return(NA_character_)
+  }
 
-  digest::digest(path, file = TRUE)
+  rlang::hash_file(path)
 }
 
 #' Capture the state of a directory.