+++ tests/testthat/test-expect-known.R
@@ -67,7 +67,6 @@ test_that("Warning for non-UTF-8 reference files", {
 
 # expect_known_value ------------------------------------------------------
 
-
 test_that("correctly matches to a file", {
   x <- 1
   expect_success(expect_known_value(x, "one.rds"))
@@ -87,7 +86,7 @@ test_that("first run is successful", {
 })
 
 test_that("equal_to_ref does not overwrite existing", {
-  tmp_rds <- tempfile(fileext=".rds")
+  tmp_rds <- tempfile(fileext = ".rds")
   on.exit(unlink(tmp_rds))
   ref_obj1 <- 1:3
   ref_obj2 <- 2:4
@@ -100,7 +99,7 @@ test_that("equal_to_ref does not overwrite existing", {
   expect_equal(readRDS(tmp_rds), ref_obj1)
 
   # Now failure does update object
-  expect_failure(expect_equal_to_reference(ref_obj2, tmp_rds, update=TRUE))
+  expect_failure(expect_equal_to_reference(ref_obj2, tmp_rds, update = TRUE))
   expect_success(expect_equal_to_reference(ref_obj2, tmp_rds))
 })