+++ tests/testthat/test-expectation.R
@@ -3,17 +3,21 @@ test_that("expectation contains failure message even when successful", {
   expect_equal(e$message, "I failed")
 })
 
-test_that("expect warns if no `failure_message`", {
-  expect_warning(expect(TRUE), "missing, with no default")
-})
-
 test_that("info only evaluated on failure", {
   expect_error(expect(TRUE, "fail", info = stop("!")), NA)
 })
 
 test_that("can subclass expectation", {
-  exp <- new_expectation("failure", "didn't work", .subclass = "foo", bar = "baz")
-  expect_true(inherits_all(exp, c("foo", "expectation_failure", "expectation", "error", "condition")))
+  exp <- new_expectation(
+    "failure",
+    "didn't work",
+    .subclass = "foo",
+    bar = "baz"
+  )
+  expect_true(inherits_all(
+    exp,
+    c("foo", "expectation_failure", "expectation", "error", "condition")
+  ))
   expect_identical(attr(exp, "bar"), "baz")
 })
 
@@ -22,7 +26,10 @@ test_that("`expect()` and `exp_signal()` signal expectations", {
   expect_error(expect(FALSE, ""), class = "expectation_failure")
 
   expect_error(exp_signal(new_expectation("success", "")), regexp = NA)
-  expect_error(exp_signal(new_expectation("failure", "")), class = "expectation_failure")
+  expect_error(
+    exp_signal(new_expectation("failure", "")),
+    class = "expectation_failure"
+  )
 })
 
 test_that("conditionMessage() is called during conversion", {