+++ tests/testthat/test-snapshot-file.R
@@ -1,5 +1,4 @@
 test_that("expect_snapshot_file works", {
-  skip_if_not(getRversion() >= "3.6.0")
   expect_snapshot_file(
     write_tmp_lines(letters),
     "foo.r",
@@ -33,7 +32,6 @@ test_that("expect_snapshot_file works", {
 
 
 test_that("expect_snapshot_file works in a different directory", {
-  skip_if_not(getRversion() >= "3.6.0")
   path <- withr::local_tempdir()
   withr::local_dir(path)
 
@@ -60,7 +58,10 @@ test_that("basic workflow", {
 
   # warns on first run
   snapper$start_file("snapshot-6", "test")
-  expect_warning(expect_snapshot_file(write_tmp_lines(letters), "letters.txt"), "Adding new")
+  expect_warning(
+    expect_snapshot_file(write_tmp_lines(letters), "letters.txt"),
+    "Adding new"
+  )
   snapper$end_file()
 
   # succeeds if unchanged
@@ -70,7 +71,10 @@ test_that("basic workflow", {
 
   # fails if changed
   snapper$start_file("snapshot-6", "test")
-  expect_failure(expect_snapshot_file(write_tmp_lines(letters[-1]), "letters.txt"))
+  expect_failure(expect_snapshot_file(
+    write_tmp_lines(letters[-1]),
+    "letters.txt"
+  ))
   snapper$end_file()
 })
 
@@ -102,10 +106,14 @@ test_that("warns on first creation", {
 
   # Errors on non-existing file
   expect_error(
-    expect_true(snapshot_file_equal(tempdir(), "test.txt", NULL, "doesnt-exist.txt")),
+    expect_true(snapshot_file_equal(
+      tempdir(),
+      "test.txt",
+      NULL,
+      "doesnt-exist.txt"
+    )),
     "`doesnt-exist.txt` not found"
   )
-
 
   # Unchanged returns TRUE
   expect_true(snapshot_file_equal(tempdir(), "test.txt", NULL, path))
@@ -158,7 +166,22 @@ test_that("snapshot_hint output differs in R CMD check", {
     testthat:::snapshot_review_hint(..., reset_output = FALSE)
   }
 
-  expect_snapshot(cat(snapshot_review_hint("lala", "foo.r", check = FALSE, ci = FALSE)))
-  expect_snapshot(cat(snapshot_review_hint("lala", "foo.r", check = TRUE, ci = FALSE)))
-  expect_snapshot(cat(snapshot_review_hint("lala", "foo.r", check = TRUE, ci = TRUE)))
+  expect_snapshot(cat(snapshot_review_hint(
+    "lala",
+    "foo.r",
+    check = FALSE,
+    ci = FALSE
+  )))
+  expect_snapshot(cat(snapshot_review_hint(
+    "lala",
+    "foo.r",
+    check = TRUE,
+    ci = FALSE
+  )))
+  expect_snapshot(cat(snapshot_review_hint(
+    "lala",
+    "foo.r",
+    check = TRUE,
+    ci = TRUE
+  )))
 })