+++ tests/testthat/test-snapshot-reporter.R
@@ -1,4 +1,3 @@
-
 test_that("can establish local snapshotter for testing", {
   snapper <- local_snapshotter()
 
@@ -13,8 +12,8 @@ test_that("basic workflow", {
   snapper <- local_snapshotter(path)
   snapper$start_file("snapshot-2")
   # output if not active (because test not set here)
-  expect_snapshot_output("x") %>%
-    expect_message("Can't save") %>%
+  expect_snapshot_output("x") |>
+    expect_message("Can't save") |>
     expect_output("[1] \"x\"", fixed = TRUE)
 
   # warns on first creation